+++ plugins/blend-importer/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog for @threepipe/blend-importer
+# Changelog for @threepipe/plugin-blend-importer
 
 All notable changes to this plugin will be documented in this file.